Scheduling and test updates.
Update the test harness to allow performance profiling.
Add tracing and scheduling parameters and use them throughout.
Current schedule focuses mostly on the 4d loop, as it is 90+% of the total runtime.
Be cleverer about iterator ordering to get the most out of vectors and threads.
Define a separate g() function for each zone.
Schedule the per-zone g() function in a way that (hopefully) eliminates redundant calls.